+++ packages/web-crawler/src/urlRules.ts
@@ -61,4 +61,21 @@ export const crawUrlRules: CrawlUrlRule[] = [
     },
     urlPattern: 'https://www.qiumiwu.com/standings/(.*)',
   },
+
+  // mozilla use jina
+  {
+    impls: ['jina'],
+    urlPattern: 'https://developer.mozilla.org(.*)',
+  },
+  // cvpr thecvf
+  {
+    impls: ['jina'],
+    urlPattern: 'https://cvpr.thecvf.com(.*)',
+  },
+  // 飞书用 jina
+  // https://github.com/lobehub/lobe-chat/issues/6879
+  {
+    impls: ['jina'],
+    urlPattern: 'https://(.*).feishu.cn/(.*)',
+  },
 ];