+++ packages/web-crawler/src/utils/htmlToMarkdown.ts
@@ -4,10 +4,28 @@ import { NodeHtmlMarkdown, type TranslatorConfigObject } from 'node-html-markdow
 
 import { FilterOptions } from '../type';
 
+const cleanObj = <T extends object>(
+  obj: T,
+): {
+  [K in keyof T as T[K] extends null ? never : K]: T[K];
+} => Object.fromEntries(Object.entries(obj).filter(([, value]) => value !== null)) as any;
+
+interface HtmlToMarkdownOutput {
+  author?: string;
+  content: string;
+  description?: string;
+  dir?: string;
+  lang?: string;
+  length?: number;
+  publishedTime?: string;
+  siteName?: string;
+  title?: string;
+}
+
 export const htmlToMarkdown = (
   html: string,
   { url, filterOptions }: { filterOptions: FilterOptions; url: string },
-) => {
+): HtmlToMarkdownOutput => {
   const window = new Window({ url });
 
   const document = window.document;
@@ -41,5 +59,17 @@ export const htmlToMarkdown = (
 
   const content = nodeHtmlMarkdown.translate(htmlNode);
 
-  return { ...parsedContent, content };
+  const result = {
+    author: parsedContent?.byline,
+    content,
+    description: parsedContent?.excerpt,
+    dir: parsedContent?.dir,
+    lang: parsedContent?.lang,
+    length: parsedContent?.length,
+    publishedTime: parsedContent?.publishedTime,
+    siteName: parsedContent?.siteName,
+    title: parsedContent?.title,
+  };
+
+  return cleanObj(result) as HtmlToMarkdownOutput;
 };