+++ src/const/settings/common.ts
@@ -2,5 +2,4 @@ import { UserGeneralConfig } from '@/types/user/settings';
 
 export const DEFAULT_COMMON_SETTINGS: UserGeneralConfig = {
   fontSize: 14,
-  themeMode: 'auto',
 };