+++ src/libs/agent-runtime/anthropic/index.ts
@@ -100,9 +100,13 @@ export class LobeAnthropicAI implements LobeRuntimeAI {
 
     const postTools = buildAnthropicTools(tools, { enabledContextCaching });
 
-    if (!!thinking) {
-      const maxTokens =
-        max_tokens ?? (thinking?.budget_tokens ? thinking?.budget_tokens + 4096 : 4096);
+    if (!!thinking && thinking.type === 'enabled') {
+      // claude 3.7 thinking has max output of 64000 tokens
+      const maxTokens = !!max_tokens
+        ? thinking?.budget_tokens && thinking?.budget_tokens > max_tokens
+          ? Math.min(thinking?.budget_tokens + max_tokens, 64_000)
+          : max_tokens
+        : 64_000;
 
       // `temperature` may only be set to 1 when thinking is enabled.
       // `top_p` must be unset when thinking is enabled.
@@ -117,7 +121,9 @@ export class LobeAnthropicAI implements LobeRuntimeAI {
     }
 
     return {
-      max_tokens: max_tokens ?? 4096,
+      // claude 3 series model hax max output token of 4096, 3.x series has 8192
+      // https://docs.anthropic.com/en/docs/about-claude/models/all-models#:~:text=200K-,Max%20output,-Normal%3A
+      max_tokens: max_tokens ?? (model.startsWith('claude-3-') ? 4096 : 8192),
       messages: postMessages,
       model,
       system: systemPrompts,